move runtime constants and types to common_runtime
