+++ pallets/amm/Cargo.toml
@@ -30,6 +30,7 @@ hydra-dx-math = {default-features = false, version = "1.0.0", git = "https://git
 # Local dependencies
 pallet-asset-registry = {path = '../asset-registry', default-features = false}
 primitives = {path = '../../primitives', default-features = false}
+common-runtime = {path = '../../runtime/common', default-features = false}
 
 # ORML dependencies
 orml-tokens = {default-features = false, version = "0.4.1-dev"}
