+++ pallets/amm/src/lib.rs
@@ -9,8 +9,8 @@ use frame_support::sp_runtime::{
 };
 use frame_support::{dispatch::DispatchResult, ensure, traits::Get, transactional};
 use frame_system::ensure_signed;
-use primitives::{asset::AssetPair, fee, traits::AMM, AssetId, Balance, Price,
-	constants::chain::{MAX_IN_RATIO, MAX_OUT_RATIO}};
+use primitives::{asset::AssetPair, fee, traits::AMM, AssetId, Balance, Price};
+use common_runtime::constants::chain::{MAX_IN_RATIO, MAX_OUT_RATIO};
 use sp_std::{marker::PhantomData, vec, vec::Vec};
 
 use frame_support::sp_runtime::app_crypto::sp_core::crypto::UncheckedFrom;
