+++ primitives/Cargo.toml
@@ -22,11 +22,11 @@ serde = {features = ['derive'], optional = true, version = '1.0.101'}
 frame-support = {default-features = false, version = '3.0.0'}
 frame-system = {default-features = false, version = '3.0.0'}
 sp-std = {default-features = false, version = '3.0.0'}
-sp-runtime = {default-features = false, version = '3.0.0'}
-sp-core = {default-features = false, version = '3.0.0'}
 
 [dev-dependencies]
+sp-core = {default-features = false, version = '3.0.0'}
 sp-io = {default-features = false, version = '3.0.0'}
+sp-runtime = {default-features = false, version = '3.0.0'}
 
 [features]
 default = ['std']
