+++ runtime/common/src/constants.rs
@@ -69,7 +69,7 @@ pub mod chain {
 	pub const MIN_TRADING_LIMIT: Balance = 1000;
 
 	pub const RUNTIME_AUTHORING_VERSION: u32 = 1;
-	pub const RUNTIME_SPEC_VERSION: u32 = 9;
+	pub const RUNTIME_SPEC_VERSION: u32 = 10;
 	pub const RUNTIME_IMPL_VERSION: u32 = 0;
 	pub const RUNTIME_TRANSACTION_VERSION: u32 = 1;
